test(backup): Move helpers to testutils
The following change will add a standalone release test generation
script that will live outside of the test directory and will use this
functions. Because of this, a better home for them is in src/testutils,
rather than in their tesat-local `__init__.py` file.

Issue: getsentry/team-ospo#159
